fix(opd8n): cleanup
